Fixed unwanted tiddler subdirectories being created.
Forward slashes are normally replaced with underscores in the `generateTiddlerFilename` method of the
FileSystemAdptor, but not when a path name filter is applied. It makes sense that file names are not
sanitised before application of path name filters because changing the characters could stop them
working as intended. However in this case forwan this case forward slashes in Pinboard bookmark descriptions
need to be replaced otherwise they are interpreted as path separators and cause unwanted subdirectories to be
created.